Pass a valid string to path_empty (#25)
From black v22.1.0 type checking is "baked in" at runtime due to the use
of mypyc. This causes a TypeError when calling path_empty, which passes
"msg=None", where a str is expected instead. To fix this behaviour, it's
easy enough to pass the same string that black itself is currently using
(see [1]).

[1]: https://github.com/psf/black/blob/f1d4e742c91dd5179d742b0db9293c4472b765f8/src/black/__init__.py#L521